% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_to_key_terms}
\alias{add_to_key_terms}
\title{Adds key terms to the key_terms.data file}
\usage{
add_to_key_terms(key_terms_to_add)
}
\arguments{
\item{key_terms_to_add}{= vector of terms to add}
}
\value{
overwrites the old rdata file with an updated version
(you can never remove existing terms, only add new ones)
}
\description{
Adds key terms to the key_terms.data file
}

