% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_function.R
\name{call_indeed_api}
\alias{call_indeed_api}
\title{Make data request to indeed.com api}
\usage{
call_indeed_api(search_term = "analyst", location = "", start_point = "",
  lim = 5)
}
\arguments{
\item{search_term}{= the term you want to search for within indeed}

\item{location}{= location of jobs to search}

\item{start_point}{= how far to look back}

\item{lim}{= max number of jobs to return.. max = 25}
}
\value{
a dataframe containing information about specific jobs that were pulled
with the api request: expired, country, state, city, company, title, href, dates, jobkey
}
\description{
Make data request to indeed.com api
}

