% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_text}
\alias{clean_text}
\title{Extract/clean text from indeed.com}
\usage{
clean_text(text = NULL, href = NULL, for_query = FALSE)
}
\arguments{
\item{href}{= link to webpage}
}
\value{
dataframe containing three columns: text, key_terms, years_exp
}
\description{
Use regular expressions to eliminate irrelevant html syntax/formatting. Also
extract keywords found in the keywords dataset and find max years experience. This can
be applied to a string of text (resume), or a href to a job posting.
}

