% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indeed_scrape.R
\name{convert_href_to_text}
\alias{convert_href_to_text}
\title{URL to job data frame function}
\usage{
convert_href_to_text(href = NULL, location = "", start_point = "")
}
\arguments{
\item{location}{= location for api}

\item{start_point}{= job number starting point for the api}

\item{href_vector}{= a vector containing hrefs}

\item{lim}{= the max number of hrefs to visit}
}
\value{
a dataframe containing href, useful text and keywords. If api is used
then it returns a dataframe containing href, useful text, keywords, and also all other api info
}
\description{
This function combines all other functions to take the href of a job posting
and output all relevant information about a job
}
