% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{cos_sim}
\alias{cos_sim}
\title{Calculates cos_sim between a resume, and job postings}
\usage{
cos_sim(resume_text_corpus, job_text_corpus)
}
\arguments{
\item{resume_text_corpus}{= corpus of the resume}

\item{job_text_corpus}{= corpus of the job postings}
}
\value{
the cosine similarity scores between
}
\description{
Calculates cos_sim between a resume, and job postings
}

