% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indeed_scrape.R
\name{coursera_scrape}
\alias{coursera_scrape}
\title{extract jobs on coursera}
\usage{
coursera_scrape(query)
}
\arguments{
\item{query}{= a string of skills separated by a plus ex. statistics+linear alg}
}
\value{
a dataframe containing href, useful text and keywords. If api is used
then it returns a dataframe containing href, useful text, keywords, and also all other api info
}
\description{
This scrapes the jobs on coursera based on what is missing from a resume
}
