% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{create_score_df}
\alias{create_score_df}
\title{Combines scoring functions to create final scores between resumes and job postings}
\usage{
create_score_df(resume_df, resume_exp = 2, job_df)
}
\arguments{
\item{resume}{= raw resume}

\item{job_db}{= a sample of the job dataaframe in posgres and new api pulls}
}
\value{
a vector: Final scores for each resume-job pair
}
\description{
Combines scoring functions to create final scores between resumes and job postings
}
