% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_text_corpus}
\alias{create_text_corpus}
\title{Uses tm_package to clean a resume or vector of documents.}
\usage{
create_text_corpus(text)
}
\arguments{
\item{text}{= a resume or vector of documents}
}
\value{
a Vcorpus/Corpus ready to turn into a document term matrix
}
\description{
Creates a corpus of the text/vector. Removes punctuation, numbers, stopwords,
whitespace, and stems the document
}

