% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_highest_degree}
\alias{extract_highest_degree}
\title{extracts highest level of degree obtained or required}
\usage{
extract_highest_degree(text)
}
\arguments{
\item{text}{=  string of text}
}
\value{
numeric value. NA = no degree, 1 = bachelors, 2 = mastrs, 3 = phd
}
\description{
extracts highest level of degree obtained or required
}

