% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_key_terms}
\alias{extract_key_terms}
\title{extracts key terms found in the text}
\usage{
extract_key_terms(text_vect, unique = TRUE)
}
\arguments{
\item{text_vect}{= vector of words}

\item{unique}{= TRUE/fALSE}

\item{string}{= TRUE/fALSE}
}
\value{
returns a string containing the keywords found separated by a space
}
\description{
extracts key terms found in the text
}
\examples{
extract_key_terms(some_vector_of_words, unique = fALSE)
}

