% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_years_experience}
\alias{extract_years_experience}
\title{extracts max number of years experience required}
\usage{
extract_years_experience(text)
}
\arguments{
\item{text}{=  string of text}
}
\value{
numeric value, the max number of years found
}
\description{
extracts max number of years experience required
}
