% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{generate_scores}
\alias{generate_scores}
\title{Combines scoring functions to create final scores between resumes and job postings}
\usage{
generate_scores(resume, href, work_exp)
}
\arguments{
\item{resume}{= raw resume}

\item{work_exp}{= work_experience of person}

\item{search_term}{= term to search job postings for}

\item{location}{= location to search job postings for}

\item{db_lim}{= limit on number of rows to query}

\item{indeed_lim}{= limit on number of indeed.com jobs to pull and score (max 25)}
}
\value{
a dataframe needed for the app
}
\description{
Combines scoring functions to create final scores between resumes and job postings
}
