% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gompertz}
\alias{gompertz}
\title{Used to determine a weight for the scores based on work experience}
\usage{
gompertz(x, a = 1, b = 8, c = 0.8)
}
\arguments{
\item{x}{= difference between years required experience and current experience}
}
\value{
numeric value. NA = no degree, 1 = bachelors, 2 = mastrs, 3 = phd
}
\description{
Used to determine a weight for the scores based on work experience
}

