% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_internship}
\alias{is_internship}
\title{Determine if posting is an internship}
\usage{
is_internship(text)
}
\arguments{
\item{text}{= text string}
}
\value{
overwrites the old rdata file with an updated version
(you can never remove existing terms, only add new ones)
}
\description{
Determine if posting is an internship
}
