% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pct_concordant}
\alias{pct_concordant}
\title{Calculate concordant pairs}
\usage{
pct_concordant(binary, score)
}
\arguments{
\item{binary}{= vector of outcomes (0 or 1) not qualified = 0 qualified = 1}
}
\value{
score = vector of scores associated with the outcomes
}
\description{
Calculate concordant pairs
}

