% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{postgres_append}
\alias{postgres_append}
\title{Hook into postgres database}
\usage{
postgres_append(new_df = NULL, append = FALSE)
}
\arguments{
\item{new_df}{= dataframe to append to current database}

\item{append}{= whether to append the database or not}
}
\value{
nothing
}
\description{
Either pull information out, or append new information
}

