% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indeed_scrape.R
\name{pull_new_job_data}
\alias{pull_new_job_data}
\title{indeed posting to data frame function}
\usage{
pull_new_job_data(search_term = "analyst", location = "",
  start_point = "", lim = 5, add_to_db = FALSE)
}
\arguments{
\item{search_term}{= search term for api}

\item{location}{= location for api}

\item{start_point}{= job number starting point for the api}

\item{lim}{= the max number of hrefs to visit}
}
\value{
a dataframe containing href, useful text and keywords. If api is used
then it returns a dataframe containing href, useful text, keywords, and also all other api info
}
\description{
This function combines specific functions that work on job postings on
indeed.com. It calls the indeed.com api for job information and extracts text off the job posting
}

