% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{query_database}
\alias{query_database}
\title{#postgresql}
\usage{
query_database(db_search_term = "", db_location = "", lim = 10000)
}
\arguments{
\item{db_search_term}{= job term to search the database for}

\item{db_location}{= location to query database}

\item{lim}{= limit the number of jobs to extract}
}
\value{
overwrites the old rdata file with an updated version
(you can never remove existing terms, only add new ones)
}
