% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{score_key_terms}
\alias{score_key_terms}
\title{Scores key terms between resumes and job postings}
\usage{
score_key_terms(job_key_terms, resume_key_terms)
}
\arguments{
\item{job_key_terms}{= a vector of strings containing key terms found within the job posting}

\item{resume_key_terms}{= a string containing key terms found within the job posting}
}
\value{
a vector: Scores for each resume-job pair
}
\description{
Scores key terms between resumes and job postings
}

