% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{update_db}
\alias{update_db}
\title{updates the postgres db by removing old jobs}
\usage{
update_db(num_days = 7, lim = 1000)
}
\arguments{
\item{num_days}{= a filter for which jobs to check.. how many days old are the jobs you want to check}

\item{lim}{= the max number of jobs to query}
}
\description{
this function identifies all jobs over a week old. It then queries the indeed api
to determine whether the job has expired or not. If it is then the row is deleted and moved into
the old_indeed_info table
}
