% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zero_range}
\alias{zero_range}
\title{Quickly determines if two numbers in the vector}
\usage{
zero_range(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{= vector of numbers}
}
\value{
TRUE if numbers are equal FALSE if they are not
}
\description{
This function is used to in the pct concordant function
}
